Format factors as `X (X%)` instead of `X% (X)`

Almost an exact copy of the equivilent function from `atable`, this function just formats factors as e.g. 15 (25 25

# S3 method for statistics_factor
format_statistics(x, factorformat = "{N} ({percent}%)", ...)

Arguments

x
a `statistics_factor` object as created by `atable:::statistics.factor`

factorformat
how to format the values

...
passed to/from other methods

Details

When `atableExtra` loads, this function automatically overrides the `atable` version

Option `factorformat` uses the `glue` package to create strings. Although `glue` parses the values within the braces (`` and ``), this is not permitted in this case as only strings are passed to `glue`. Values that can be used in the are braces are `N`, `total` and `percent`.

Examples

data(mtcars)
mtcars$cyl <- as.factor(mtcars$cyl)
atable(mtcars, target_cols = "cyl")

#>   Group         value   
#> 1 Observations          
#> 2               32      
#> 3 cyl                   
#> 4      4        11 (34%)
#> 5      6        7 (22%) 
#> 6      8        14 (44%)
#> 7      missing  0 (0%)