Format summary statistics for numeric variables when using numstats

format_statistics.numeric_stats can use any statistic returned by numstats. It returns the type of table expected by atable.

# S3 method for numeric_stats
format_statistics(
  x,
  numstats = c(`Mean (SD)` = "{mean} ({sd})"),
  missingformat = c(`Valid (missing)` = "{Nvalid} ({Nmissing})"),
  ...
)

Arguments

x
a numstats object

numstats
either a number (1 to 3) or a named character, where the name is used as the row name and the second part in the formatting. See details for more info.

missingformat
either a logical or a named string defining the format as for numstats.

...
passed to/from other methods

Details

Options numstats and missingformat use the glue package to create strings. Although glue parses the values within the braces (`{` and `}`), this is not permitted in this case as only strings are passed to glue after being formatted by atable_options("format_numbers"). To format the numbers differently, reassign that function.

The numstats argument allows the flexible formatting of the statistics and label. It accepts either a number (1 to 3) or a names character string. Number 1 is equivalent to c("Mean (SD)" = "{mean} ({sd})"). Number 2 is equivalent to c("Min - Max" = "{q0} - {q1}") and number 3 is equivalent to c("Median [Quartiles]" = "{q0.5} [{q0.25}; {q0.75}]"). Any parameter returned by numeric_stats can be used.

missingformat works in the same way as numstats. Tt defaults to c("Valid (missing)" = "{Nvalid} ({Nmissing})").

Examples

require(atable)
atable_options(statistics.numeric = numeric_stats)
data(mtcars)
# atableExtra default
atable(mpg ~ am, mtcars)

#> Warning: cannot compute exact p-value with ties

#>                    Group        0        1      p stat   Effect Size (CI)
#> 1          Observations                                                  
#> 2          \\quad \\quad       19       13                               
#> 3             mpg \\quad                                                 
#> 4       \\quad Mean (SD) 17 (3.8) 24 (6.2) 0.0039 0.64 -1.5 (-2.3; -0.65)
#> 5 \\quad Valid (missing)   19 (0)   13 (0)                               

# specify only mean
atable(mpg ~ am, mtcars, numstats = c("Mean (SD)" = "{mean} ({sd})"))

#> Warning: cannot compute exact p-value with ties

#>                    Group        0        1      p stat   Effect Size (CI)
#> 1          Observations                                                  
#> 2          \\quad \\quad       19       13                               
#> 3             mpg \\quad                                                 
#> 4       \\quad Mean (SD) 17 (3.8) 24 (6.2) 0.0039 0.64 -1.5 (-2.3; -0.65)
#> 5 \\quad Valid (missing)   19 (0)   13 (0)                               

# specify options by number
atable(mpg ~ am, mtcars, numstats = 1:2)

#> Warning: cannot compute exact p-value with ties

#>                    Group        0        1      p stat   Effect Size (CI)
#> 1          Observations                                                  
#> 2          \\quad \\quad       19       13                               
#> 3             mpg \\quad                                                 
#> 4       \\quad Mean (SD) 17 (3.8) 24 (6.2) 0.0039 0.64 -1.5 (-2.3; -0.65)
#> 5       \\quad Min - Max  10 - 24  15 - 34                               
#> 6 \\quad Valid (missing)   19 (0)   13 (0)                               

atable(mpg ~ am, mtcars, numstats = c(1,3))

#> Warning: cannot compute exact p-value with ties

#>                       Group           0           1      p stat
#> 1             Observations                                     
#> 2             \\quad \\quad          19          13            
#> 3                mpg \\quad                                    
#> 4          \\quad Mean (SD)    17 (3.8)    24 (6.2) 0.0039 0.64
#> 5 \\quad Median [Quartiles] 17 [15; 19] 23 [21; 30]            
#> 6    \\quad Valid (missing)      19 (0)      13 (0)            
#>     Effect Size (CI)
#> 1                   
#> 2                   
#> 3                   
#> 4 -1.5 (-2.3; -0.65)
#> 5                   
#> 6                   

# change to square brackets
atable(mpg ~ am, mtcars, numstats = c("Mean [SD]" = "{mean} [{sd}]"))

#> Warning: cannot compute exact p-value with ties

#>                    Group        0        1      p stat   Effect Size (CI)
#> 1          Observations                                                  
#> 2          \\quad \\quad       19       13                               
#> 3             mpg \\quad                                                 
#> 4       \\quad Mean [SD] 17 [3.8] 24 [6.2] 0.0039 0.64 -1.5 (-2.3; -0.65)
#> 5 \\quad Valid (missing)   19 (0)   13 (0)